Implement CPU, GPU and memory gauges
Every two seconds, CPU and memory usage will be polled and the gauges
will be redrawn based on that.

The GPU gauge is currently disabled, as polling GPU usage has not been
implemented yet. Gigabyte's implementation for their Control Center
polls the usage of the dedicated GPU, which is what we want to do.