Add initial embedded controller support
This code is currently disabled, as it will only work if the entire
program is running as root and the ec_sys kernel module is loaded.

One solution would be to just run the program as root, but I want to
avoid that.

Another solution would be to make a new kernel driver that can expose
certain embedded controller (EC) registers via hwmon or sysfs, and then
have our application interact with that driver. Although more difficult,
it is safer because we are not making the application probe random EC
registers and potentially breaking the system.