Switch to libusb
Although hidapi is able to connect to certain interfaces, the setup data
is still wrong. For whatever reason, "wIndex" is set to 0 instead of 3,
and there is simply no way to fix that with hidapi. This is a problem,
as changing the RGB requires specific setup data sent to the keyboard.

Because of this, libusb will be used instead. In my test program, it was
able to change the RGB on my keyboard, whereas hidapi could not.