Add custom graphics view to UI
The events in the default QGraphicsView class do nothing, which means
there is no interaction between the view and the scene. Because of this,
we will need to create a subclass of it and override the events.

While we do this, move RGB setup code into its own function in
MainWindow, and set the current tab to the main one every time the
application starts.