RGBKeyboard: Support attaching light bar
We can now interact with the light bar, which seems to behave the
same way as the keyboard for RGB. However, should it be present,
getting the feature report for the keyboard requires setting the
reserved value to 0x02 or else we don't get any data.

While adjusting RGB settings for the light bar can be implemented
now, it will be useless without UI changes, which has not taken
place yet.